Implemented hashing password

+++ db.py
@@ -21,10 +21,15 @@ class DB:
         return self.accounts_collection.count_documents({"username": username}) > 0
 
     def register(self, username, password):
+        if self.is_account_exist(username):
+            print("Account already exists")
+            return False
+
         salt = secrets.token_hex(16)
         hashed_password = self._hash_password(password, salt)
         account = {"username": username, "password": hashed_password, "salt": salt}
         self.accounts_collection.insert_one(account)
+        return True
 
     def _hash_password(self, password, salt):
         return hashlib.sha256((password + salt).encode("utf-8")).hexdigest()
@@ -41,10 +46,19 @@ class DB:
         return self.online_peers_collection.count_documents({"username": username}) > 0
 
     def user_login(self, username, ip, port):
+        if self.is_account_online(username):
+            print("Account is already online")
+            return False
+
         online_peer = {"username": username, "ip": ip, "port": port}
         self.online_peers_collection.insert_one(online_peer)
+        return True
 
     def user_logout(self, username):
+        if not self.is_account_online(username):
+            print("Account is not online")
+            return False
+
         self.online_peers_collection.delete_one({"username": username})
 
     def get_peer_ip_port(self, username):
@@ -54,3 +68,15 @@ class DB:
         else:
             return None
 
+    def drop_all_records(self):
+        self.accounts_collection.delete_many({})
+        self.online_peers_collection.delete_many({})
+
+
+# Dummy Code to Test Database
+# if __name__ == "__main__":
+#     db = DB()
+
+#     # db.register("Ayman", "False")
+#     if db.verify_password("Ayman", "False"):
+#         print("Password is correct")
